(tae): ut code refactor 2
