+++ pkg/vm/engine/tae/db/compound_test.go
@@ -9,40 +9,8 @@ import (
 	"github.com/matrixorigin/matrixone/pkg/vm/engine/tae/iface/data"
 	"github.com/matrixorigin/matrixone/pkg/vm/engine/tae/iface/handle"
 	"github.com/matrixorigin/matrixone/pkg/vm/engine/tae/model"
-	"github.com/matrixorigin/matrixone/pkg/vm/engine/tae/tables/jobs"
 	"github.com/stretchr/testify/assert"
 )
-
-func compactBlocks(t *testing.T, e *DB, dbName string, schema *catalog.Schema, skipConflict bool) {
-	txn, _ := e.StartTxn(nil)
-	db, _ := txn.GetDatabase(dbName)
-	rel, _ := db.GetRelationByName(schema.Name)
-
-	var metas []*catalog.BlockEntry
-	it := rel.MakeBlockIt()
-	for it.Valid() {
-		blk := it.GetBlock()
-		if blk.Rows() < int(schema.BlockMaxRows) {
-			it.Next()
-			continue
-		}
-		meta := blk.GetMeta().(*catalog.BlockEntry)
-		metas = append(metas, meta)
-		it.Next()
-	}
-	for _, meta := range metas {
-		task, err := jobs.NewCompactBlockTask(nil, txn, meta, e.Scheduler)
-		assert.NoError(t, err)
-		err = task.OnExec()
-		if !skipConflict {
-			assert.NoError(t, err)
-		}
-	}
-	err := txn.Commit()
-	if !skipConflict {
-		assert.NoError(t, err)
-	}
-}
 
 func TestCompoundPK1(t *testing.T) {
 	tae := initDB(t, nil)
