+++ pkg/vm/engine/tae/db/gc_test.go
@@ -15,112 +15,19 @@
 package db
 
 import (
-	"io/ioutil"
-	"strconv"
-	"strings"
 	"sync"
 	"testing"
 	"time"
 
-	gbat "github.com/matrixorigin/matrixone/pkg/container/batch"
 	"github.com/matrixorigin/matrixone/pkg/vm/engine/tae/catalog"
 	"github.com/matrixorigin/matrixone/pkg/vm/engine/tae/common"
 	"github.com/matrixorigin/matrixone/pkg/vm/engine/tae/container/compute"
-	"github.com/matrixorigin/matrixone/pkg/vm/engine/tae/options"
 	"github.com/matrixorigin/matrixone/pkg/vm/engine/tae/tables/jobs"
 	"github.com/matrixorigin/matrixone/pkg/vm/engine/tae/testutils"
+	"github.com/matrixorigin/matrixone/pkg/vm/engine/tae/testutils/config"
 	"github.com/panjf2000/ants/v2"
 	"github.com/stretchr/testify/assert"
 )
-
-func printCheckpointStats(t *testing.T, tae *DB) {
-	t.Logf("GetCheckpointedLSN: %d", tae.Wal.GetCheckpointed())
-	t.Logf("GetPenddingLSNCnt: %d", tae.Wal.GetPenddingCnt())
-}
-
-func getSegmentFileNames(dir string) (names map[uint64]string) {
-	names = make(map[uint64]string)
-	files, err := ioutil.ReadDir(dir)
-	if err != nil {
-		panic(err)
-	}
-	for _, f := range files {
-		name := f.Name()
-		segName := strings.TrimSuffix(name, ".seg")
-		if segName == name {
-			continue
-		}
-		id, err := strconv.ParseUint(segName, 10, 64)
-		if err != nil {
-			panic(err)
-		}
-		names[id] = name
-	}
-	return
-}
-
-func enableCheckpoint(in *options.Options) (opts *options.Options) {
-	if in == nil {
-		opts = new(options.Options)
-	} else {
-		opts = in
-	}
-	opts.CheckpointCfg = new(options.CheckpointCfg)
-	opts.CheckpointCfg.ScannerInterval = 10
-	opts.CheckpointCfg.ExecutionLevels = 5
-	opts.CheckpointCfg.ExecutionInterval = 1
-	opts.CheckpointCfg.CatalogCkpInterval = 5
-	opts.CheckpointCfg.CatalogUnCkpLimit = 1
-	return opts
-}
-
-func appendFailClosure(t *testing.T, data *gbat.Batch, name string, e *DB, wg *sync.WaitGroup) func() {
-	return func() {
-		if wg != nil {
-			defer wg.Done()
-		}
-		txn, _ := e.StartTxn(nil)
-		database, _ := txn.GetDatabase("db")
-		rel, _ := database.GetRelationByName(name)
-		err := rel.Append(data)
-		assert.NotNil(t, err)
-		assert.Nil(t, txn.Rollback())
-	}
-}
-
-func appendClosure(t *testing.T, data *gbat.Batch, name string, e *DB, wg *sync.WaitGroup) func() {
-	return func() {
-		if wg != nil {
-			defer wg.Done()
-		}
-		txn, _ := e.StartTxn(nil)
-		database, _ := txn.GetDatabase("db")
-		rel, _ := database.GetRelationByName(name)
-		err := rel.Append(data)
-		assert.Nil(t, err)
-		assert.Nil(t, txn.Commit())
-	}
-}
-
-func tryAppendClosure(t *testing.T, data *gbat.Batch, name string, e *DB, wg *sync.WaitGroup) func() {
-	return func() {
-		if wg != nil {
-			defer wg.Done()
-		}
-		txn, _ := e.StartTxn(nil)
-		database, _ := txn.GetDatabase("db")
-		rel, err := database.GetRelationByName(name)
-		if err != nil {
-			_ = txn.Rollback()
-			return
-		}
-		if err = rel.Append(data); err != nil {
-			_ = txn.Rollback()
-			return
-		}
-		_ = txn.Commit()
-	}
-}
 
 func TestGCBlock1(t *testing.T) {
 	tae := initDB(t, nil)
@@ -130,16 +37,11 @@ func TestGCBlock1(t *testing.T) {
 	schema.SegmentMaxBlocks = 2
 
 	bat := catalog.MockData(schema, schema.BlockMaxRows)
-	txn, _ := tae.StartTxn(nil)
-	db, _ := txn.CreateDatabase("db")
-	rel, _ := db.CreateRelation(schema)
-	err := rel.Append(bat)
-	assert.Nil(t, err)
-	assert.Nil(t, txn.Commit())
+	createRelationAndAppend(t, tae, "db", schema, bat, true)
 
-	txn, _ = tae.StartTxn(nil)
-	db, _ = txn.GetDatabase("db")
-	rel, _ = db.GetRelationByName(schema.Name)
+	txn, _ := tae.StartTxn(nil)
+	db, _ := txn.GetDatabase("db")
+	rel, _ := db.GetRelationByName(schema.Name)
 	it := rel.MakeBlockIt()
 	blk := it.GetBlock()
 	meta := blk.GetMeta().(*catalog.BlockEntry)
@@ -165,7 +67,7 @@ func TestGCBlock1(t *testing.T) {
 }
 
 func TestAutoGC1(t *testing.T) {
-	opts := enableCheckpoint(nil)
+	opts := config.WithQuickScanAndCKPOpts(nil)
 	tae := initDB(t, opts)
 	defer tae.Close()
 	schema := catalog.MockSchemaAll(13, 3)
@@ -175,14 +77,8 @@ func TestAutoGC1(t *testing.T) {
 	totalRows := schema.BlockMaxRows * 21 / 2
 	bat := catalog.MockData(schema, totalRows)
 	bats := compute.SplitBatch(bat, 100)
+	createRelation(t, tae, "db", schema, true)
 	pool, _ := ants.NewPool(50)
-	{
-		txn, _ := tae.StartTxn(nil)
-		database, _ := txn.CreateDatabase("db")
-		_, err := database.CreateRelation(schema)
-		assert.Nil(t, err)
-		assert.Nil(t, txn.Commit())
-	}
 	var wg sync.WaitGroup
 	for _, data := range bats {
 		wg.Add(1)
@@ -216,7 +112,7 @@ func TestAutoGC1(t *testing.T) {
 // 3. Create a table w one appendable block data and commit
 // 4. Drop the table and commit
 func TestGCTable(t *testing.T) {
-	opts := enableCheckpoint(nil)
+	opts := config.WithQuickScanAndCKPOpts(nil)
 	tae := initDB(t, opts)
 	defer tae.Close()
 	schema := catalog.MockSchemaAll(13, 12)
@@ -224,17 +120,10 @@ func TestGCTable(t *testing.T) {
 	schema.SegmentMaxBlocks = 2
 
 	// 1. Create a table without data
-	txn, _ := tae.StartTxn(nil)
-	db, _ := txn.CreateDatabase("db")
-	_, _ = db.CreateRelation(schema)
-	assert.NoError(t, txn.Commit())
+	db, _ := createRelation(t, tae, "db", schema, true)
 
 	// 2. Drop the table
-	txn, _ = tae.StartTxn(nil)
-	db, _ = txn.GetDatabase("db")
-	_, err := db.DropRelationByName(schema.Name)
-	assert.NoError(t, err)
-	assert.NoError(t, txn.Commit())
+	dropRelation(t, tae, "db", schema.Name)
 
 	dbEntry, _ := tae.Catalog.GetDatabaseByID(db.GetID())
 	now := time.Now()
@@ -249,22 +138,13 @@ func TestGCTable(t *testing.T) {
 	bats := compute.SplitBatch(bat, 4)
 
 	// 3. Create a table and append 7 rows
-	txn, _ = tae.StartTxn(nil)
-	db, _ = txn.GetDatabase("db")
-	rel, _ := db.CreateRelation(schema)
-	err = rel.Append(bats[0])
-	assert.NoError(t, err)
-	assert.NoError(t, txn.Commit())
+	db, _ = createRelationAndAppend(t, tae, "db", schema, bats[0], false)
 
 	names := getSegmentFileNames(tae.Dir)
 	assert.Equal(t, 1, len(names))
 
 	// 4. Drop the table
-	txn, _ = tae.StartTxn(nil)
-	db, _ = txn.GetDatabase("db")
-	_, err = db.DropRelationByName(schema.Name)
-	assert.NoError(t, err)
-	assert.NoError(t, txn.Commit())
+	dropRelation(t, tae, "db", schema.Name)
 
 	dbEntry, _ = tae.Catalog.GetDatabaseByID(db.GetID())
 	now = time.Now()
@@ -278,13 +158,7 @@ func TestGCTable(t *testing.T) {
 	assert.Equal(t, 0, len(names))
 
 	// 5. Create a table and append 3 block
-	txn, _ = tae.StartTxn(nil)
-	db, _ = txn.GetDatabase("db")
-	rel, err = db.CreateRelation(schema)
-	assert.NoError(t, err)
-	err = rel.Append(bat)
-	assert.NoError(t, err)
-	assert.NoError(t, txn.Commit())
+	db, _ = createRelationAndAppend(t, tae, "db", schema, bat, false)
 	names = getSegmentFileNames(tae.Dir)
 	t.Log(names)
 	assert.Equal(t, 2, len(names))
@@ -293,11 +167,7 @@ func TestGCTable(t *testing.T) {
 	compactBlocks(t, tae, "db", schema, true)
 
 	// 6. Drop the table
-	txn, _ = tae.StartTxn(nil)
-	db, _ = txn.GetDatabase("db")
-	_, err = db.DropRelationByName(schema.Name)
-	assert.NoError(t, err)
-	assert.NoError(t, txn.Commit())
+	dropRelation(t, tae, "db", schema.Name)
 	testutils.WaitExpect(200, func() bool {
 		return dbEntry.CoarseTableCnt() == 0
 	})
@@ -309,11 +179,7 @@ func TestGCTable(t *testing.T) {
 	assert.Equal(t, 0, len(names))
 
 	// 7. Create a table
-	txn, _ = tae.StartTxn(nil)
-	db, _ = txn.GetDatabase("db")
-	_, err = db.CreateRelation(schema)
-	assert.NoError(t, err)
-	assert.NoError(t, txn.Commit())
+	db, _ = createRelation(t, tae, "db", schema, false)
 
 	// 8. Append blocks and drop
 	var wg sync.WaitGroup
@@ -322,20 +188,16 @@ func TestGCTable(t *testing.T) {
 	bats = compute.SplitBatch(bat, 20)
 	for i := range bats[:10] {
 		wg.Add(1)
-		pool.Submit(tryAppendClosure(t, bats[i], schema.Name, tae, &wg))
+		_ = pool.Submit(tryAppendClosure(t, bats[i], schema.Name, tae, &wg))
 	}
 	wg.Add(1)
-	pool.Submit(func() {
+	_ = pool.Submit(func() {
 		defer wg.Done()
-		txn, _ := tae.StartTxn(nil)
-		db, _ := txn.GetDatabase("db")
-		_, err := db.DropRelationByName(schema.Name)
-		assert.NoError(t, err)
-		assert.NoError(t, txn.Commit())
+		dropRelation(t, tae, "db", schema.Name)
 	})
 	for i := range bats[10:] {
 		wg.Add(1)
-		pool.Submit(tryAppendClosure(t, bats[i+10], schema.Name, tae, &wg))
+		_ = pool.Submit(tryAppendClosure(t, bats[i+10], schema.Name, tae, &wg))
 	}
 	wg.Wait()
 	printCheckpointStats(t, tae)
