+++ pkg/vm/engine/tae/db/replay_test.go
@@ -10,10 +10,6 @@ import (
 	"github.com/matrixorigin/matrixone/pkg/vm/engine/tae/common"
 	"github.com/matrixorigin/matrixone/pkg/vm/engine/tae/container/compute"
 	"github.com/matrixorigin/matrixone/pkg/vm/engine/tae/iface/handle"
-
-	// "github.com/matrixorigin/matrixone/pkg/vm/engine/tae/tables/jobs"
-	// "github.com/matrixorigin/matrixone/pkg/vm/engine/tae/tasks"
-	// ops "github.com/matrixorigin/matrixone/pkg/vm/engine/tae/tasks/worker"
 	"github.com/matrixorigin/matrixone/pkg/vm/engine/tae/testutils"
 	"github.com/panjf2000/ants/v2"
 	"github.com/stretchr/testify/assert"
